requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764738024/work
certifi==2022.12.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179421243/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1681508727236/work
discord.py==2.2.2
frozenlist==1.3.3
github3.py==4.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
multidict==6.0.4
openai==0.27.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT==2.6.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1680286922386/work
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
yarl==1.9.2
arxiv
praw
